# docs/swagger.yaml
basePath: /health
definitions:
...
  User:
    properties:
      consent:
        type: boolean
      encrypted_blob:
        type: string
      encrypted_key:
        type: string
      exposure_notification:
        type: boolean
      id:
        type: string
      public_key:
        type: string
      re_post:
        type: boolean
      uuid:
        type: string
    type: object
...
host: localhost
info:
  contact: {}
  description: Rokwire Health Building Block API Documentation.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Rokwire Health Building Block API
  version: 1.9.0
paths:
...
  /covid19/user:
    get:
      consumes:
      - application/json
      description: Gives the current user.
      operationId: getUser
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/User'
      security:
      - AppUserAuth: []
      tags:
      - Covid19
...
schemes:
- https
securityDefinitions:
  AdminUserAuth:
    in: header (add Bearer prefix to the Authorization value)
    name: Authorization
    type: apiKey
  AppUserAuth:
    in: header (add Bearer prefix to the Authorization value)
    name: Authorization
    type: apiKey
  ProvidersAuth:
    in: header
    name: ROKWIRE-HS-API-KEY
    type: apiKey
  RokwireAuth:
    in: header
    name: ROKWIRE-API-KEY
    type: apiKey
swagger: "2.0"